+++ src/adhocracy_meinberlin/adhocracy_meinberlin/resources/bplan.py
@@ -1,8 +1,11 @@
 """BPlan process resources."""
+from pyramid.interfaces import IRequest
+from adhocracy_core.interfaces import IResource
 from adhocracy_core.resources import add_resource_type_to_registry
 from adhocracy_core.resources import process
 from adhocracy_core.resources import proposal
-import adhocracy_core.sheets.embed
+from adhocracy_core.sheets.embed import embed_code_config_adapter
+from adhocracy_core.sheets.embed import IEmbedCodeConfig
 import adhocracy_meinberlin.sheets.bplan
 import adhocracy_core.sheets.image
 
@@ -49,8 +52,22 @@ process_meta = process.process_meta._replace(
 )
 
 
+def embed_code_config_bplan_adapter(context: IResource,
+                                    request: IRequest) -> {}:
+    """Return config to render `adhocracy_core:templates/embed_code.html`."""
+    mapping = embed_code_config_adapter(context, request)
+    mapping.update({'widget': 'mein-berlin-bplaene-proposal-embed',
+                    'style': 'height: 650px',
+                    'noheader': 'true',
+                    })
+    return mapping
+
+
 def includeme(config):
     """Add resource type to content."""
     add_resource_type_to_registry(proposal_meta, config)
     add_resource_type_to_registry(proposal_version_meta, config)
     add_resource_type_to_registry(process_meta, config)
+    config.registry.registerAdapter(embed_code_config_bplan_adapter,
+                                    (IProcess, IRequest),
+                                    IEmbedCodeConfig)
