+++ src/adhocracy_meinberlin/adhocracy_meinberlin/resources/test_bplan.py
@@ -1,3 +1,4 @@
+from pyramid import testing
 from pytest import mark
 from pytest import fixture
 
@@ -73,3 +74,24 @@ class TestProcess:
         assert registry.content.create(meta.iresource.__identifier__)
 
 
+class TestEmbedCodeConfigAdapter:
+
+    @mark.usefixtures('integration')
+    def test_get_config_for_bplan_process(self, request_, registry):
+        from adhocracy_core.sheets.embed import IEmbedCodeConfig
+        from .bplan import IProcess
+        context = testing.DummyResource(__provides__=IProcess)
+        result = registry.getMultiAdapter((context, request_),
+                                          IEmbedCodeConfig)
+        assert result == {'sdk_url': 'http://localhost:6551/AdhocracySDK.js',
+                          'frontend_url': 'http://localhost:6551',
+                          'path': 'http://example.com/',
+                          'widget': 'mein-berlin-bplaene-proposal-embed',
+                          'autoresize': 'false',
+                          'locale': 'en',
+                          'autourl': 'false',
+                          'initial_url': '',
+                          'nocenter': 'true',
+                          'noheader': 'true',
+                          'style': 'height: 650px',
+                          }
