+++ src/adhocracy_meinberlin/adhocracy_meinberlin/sheets/__init__.py
@@ -6,4 +6,3 @@ def includeme(config):  # pragma: no cover
     config.include('.kiezkassen')
     config.include('.bplan')
     config.include('.burgerhaushalt')
-    config.include('.embed')
